rename ingest to ingest_message
