+++ app/src/util.py
@@ -3,11 +3,11 @@ from collections.abc import Callable
 
 import requests
 
-import ingest
+import ingest_message
 
 
 def exists_in_cmr(
-    cmr_domain: str, short_name: ingest.Collection, granule_ur: str, granule_ur_pattern: Callable[[str], str]
+    cmr_domain: str, short_name: ingest_message.Collection, granule_ur: str, granule_ur_pattern: Callable[[str], str]
 ) -> bool:
     url = f'https://{cmr_domain}/search/granules.umm_json'
     params = {
